Fix GPU high-perf parsing bug found by tibeer.
Tim found that flavor-name-check.py would not accept SCS-2V:4-GI1h.
The reason is that the parser interprets this as generation 1h rather
than gen1 with high-performance, and gen 1h is undefined.

Interpreting the spec strictly, the tool was right: We can only
specify GPU high-performance after a CU/EU/SM count, meaning that
-GI1:20h is legal, but -GI1h and -GI1:h are not.
The spec currentyly says -GX[N][:M[h]] or -gX[N][:M[h]].
The unfortunate truth is that the tool does allow the generation
of such illegal names in -i (interactive input) mode.

Hint: Use -v to see how the tool interprets all the suffices.
-GI1h means pass-through intel GPU gen Xe1 with high frequency
(but unspecified number of EUs).

Two options to resolve this:
(A) We prevent the generation of -GI1h in -i mode, preventing a
   high-frequency specification without specifying the number of
   accessible CU/EU/SMs.
(B) We change the parser to accept -GI1h. We should then also
   clarify the specification as it says -GX[N][:M[h]] and should
   instead then say -GX[N][:M[h]] OR -GX[N][h]

This patch implements the tool fix for approach (B).
The doc change will be in a second commit.

Signed-off-by: Kurt Garloff <kurt@garloff.de>