Feat/flavor checker direct (#163)
* Add script using OpenStack API to check flavors.

We get the list of flavors and check compliance of them.

* Print output report in YAML.

* Fix output report dict structure.

* Check vCPUs, RAM, Disk OpenStack data vs name.

Overproviding is allowed, we only warn.
For underprovisioning, we produce an error.

* Count errors for underperforming flavors.

* Some comments help with code reading.

* Warn for strange memory sizes.

We only recommend 0.5, 1, 1.5, 2, ... 9.5, 10, 11, 12, 13, ...
and warn for numbers that are not integers if >= 10 and not halfed
integers for smaller memory sizes. (This gives us still plenty of
flexibility.)

Another change was naming the return values as cpuram is more readable
than ret[0].

* Fixed formatting.

- Replaced tabs by spaces
- switched partly to newer string formatting style
- fixed indentation issues
- removed code-linebreaks as they reduce readability
- removed trailing backslashes inside braces as they are obsolete

* Implement Tim's suggestion who to report result.

Tim suggested a better report format discussing PR#163 at
https://github.com/SovereignCloudStack/Docs/pull/163#issuecomment-1378360323
which we implement.

Also: Don't list all flavors by category by default, but just the
Summary with the numbers. Option -v/--verbose produces the full
report though.

Sidenote: As we have dicts, the field order is determined by the Python
implementation (using Alphabetical ordering). If we prefer a different
ordering, we might need to use some ordered dict implementation.

* Make it pass flake8 tests.

* Address a few of the pylint complaints.

Mainly superfluous () after if. (Yes, I have written C for too long.)

Signed-off-by: Kurt Garloff <kurt@garloff.de>

Signed-off-by: Kurt Garloff <kurt@garloff.de>
Signed-off-by: Tim Beermann <beermann@osism.tech>
Co-authored-by: Tim Beermann <beermann@osism.tech>