SCS compliance check tool PoC.
* It parses the passed spec YAML file
* It identifies the valid standards by date
* And checks all wanted layers
* By running the tests.

TODO: Allow passing -q/--quiet, -v/--verbose.
TODO: Allow passing -d/--date to check life at a different point in
time.
TODO: We have changed the spec.yaml format somewhat, as we did not
want to manually specify (and then adjust based on time) stable and
upcoming version. This needs to be reflected in the spec.

This is the output from 2hrs pair programming Eduard/Kurt.

Signed-off-by: Kurt Garloff <kurt@garloff.de>