Only two layers for now. Updated date handling. Tooling.
We had decided that we only have tow layers for now.

Explicitly listing which version is stable and which one is upcoming
is unncessary. Worse, it requires us to do commits to the repo
just because the calendar advanced even if nothing else changed.
Change the design to have each version (of a standards collection)
have an effective date (stabilizied_at) and optionally an expiration
date (replaced_at).

Update docs to reflect this.

Also add pointer to tooling.

Signed-off-by: Kurt Garloff <kurt@garloff.de>