Create copy.deepcopy() before adding errors.
This avoids the errors dict key to be added to the original version
spec. (Adding it would create a warning from an unknown keyword errors.)

This will make the version_index = report[...].index(bestversion) fail.
As we don't have bestversion in report, but a deep copy, that is to be
expected.

We fortunately don't need to determine the index, it's always 0 (as we
only copy this one).

Signed-off-by: Kurt Garloff <kurt@garloff.de>