Stabilize the SSD flavor decision.
Note: This means that our plan to implement these and require these at
a later point in time in our list of mandatory flavors in the
flavor-naming-spec.md. The precise time when these will become
required is to be determined in a separate discussion.
(My current assumption is shortly after R4, i.e. April 2023.)

Signed-off-by: Kurt Garloff <kurt@garloff.de>