Parse and validate new v2 flavor names.
We have changed the naming scheme to avoid ':' in flavor names as
k8s does not like these within a label.
Ignoring the SCS- prefix (which stays the same, we did not add version
and leave the trailing dash as is), we transform:
* v1 '-' becomes v2 '_'
* v1 ':' becomes v2 '-'

The flavor-name-check.py script still understands the old names unless
you pass the "-2" option, which results in old flavor names no longer
being acceptable.

The old names however don't count towards the mandatory flavor list.
So a cloud that offers new and old names would be compliant without
-2, but only the old flavors are insufficient.

Caveats:
* We probably should have a -1 option to allow only old flavors
  (and count them towards the new mandatory list)
* Without -2, we are too liberal in accepting mixed flavor names:
  We accept flavors that start with the v2 scheme and then switch
  to v1. (This is not considered an urgent issue right now, as we
  expect the transition time to be limited.)

Signed-off-by: Kurt Garloff <kurt@garloff.de>