Produce input for flavor manager. (#227)
* Produce input for flavor manager.

We don't want to duplicate information in our primary file
SCS-Spec.MandatoryFlavors.yaml. Instead we generate the input in the
wanted format for the OpenStack flavor-manager.

Addresses #216.

* Never turst pylint's advice ...
* Avoid sorting yaml dicts as suggested by @itrich.
* Generate alias fields in the description.

If we output the new names (i.e. option -1 is not specified), we
also put an alias in the description field. This may or may not be
useful for the transition.

Signed-off-by: Kurt Garloff <kurt@garloff.de>