Add option -o to accept v1 names fulfilling v2 mandatory flavors.
While our test without option -2 (--v2plus) would still have tolerated
the old v1 flavors, it would not count them as fulfilling the list
of mandatory flavors. With the new option -o/--accept-old-mand,
this is now allowed.

Alternative would have been to read the list of mandatory flavors
from an old file with v1 names. We would have had to do this, if we
had added the two new SSD flavors to v2. We have not yet :-(

Signed-off-by: Kurt Garloff <kurt@garloff.de>