v2 testing does not tolerate v1 mand flavors for v2.
We have determined that v2 compliance requires v2 mandatory flavor
names and do both v1 and v2 compliance testing and expose both results
on the github page. This makes the unfortunate (minor) incompatibility
visible, which is better.

So the new `--accept-old-mand` flag is not used in compliance testing
normally. Having it is still useful ...

Signed-off-by: Kurt Garloff <kurt@garloff.de>