Fix/accept v1 mandatory flavors (#260)
* Add option -o to accept v1 names fulfilling v2 mandatory flavors.

While our test without option -2 (--v2plus) would still have tolerated
the old v1 flavors, it would not count them as fulfilling the list
of mandatory flavors. With the new option -o/--accept-old-mand,
this is now allowed.

Alternative would have been to read the list of mandatory flavors
from an old file with v1 names. We would have had to do this, if we
had added the two new SSD flavors to v2. We have not yet :-(

* Use new --accept-old-mand option to allow v1 clouds.

* v2 testing does not tolerate v1 mand flavors for v2.

We have determined that v2 compliance requires v2 mandatory flavor
names and do both v1 and v2 compliance testing and expose both results
on the github page. This makes the unfortunate (minor) incompatibility
visible, which is better.

So the new `--accept-old-mand` flag is not used in compliance testing
normally. Having it is still useful ...

* Remove (commented out) debugging code.

Signed-off-by: Kurt Garloff <kurt@garloff.de>