Refcator the document to make it more readable.
Main change is to move the complex and rarely used extensions
to the bottom of the document and add a section that explains why
these exist at all and how they may or may not be used.

The change does not change the specification in any way, thus no
new version number is assigned.

This change is in preparation of two decsisions and PRs that will
be created:
(1) Reduce the mandated flavors by making the flavors with disks
    only recommended (as discussed on May 8).
(2) Discuss an alternative approach to flavor naming from PS.

Signed-off-by: Kurt Garloff <kurt@garloff.de>