Introduce the SSD flavors as mandatory flavors.
Note: We always intended these to be mandatory. This could be
reconsidered to avoid mandatory flavors with root disks completely.
The reason not to do so: Local SSD disks are not the default storage
class and it will be difficult to specify them via
block-device-mapping-v2 at VM creation time. So to make them useful
at all, we should have a root disk included.

Signed-off-by: Kurt Garloff <kurt@garloff.de>