Decisions also live in the Standards directory.
While it was correct to change the scs-0110-v1 SSD type to a decision
(this is not a standard on its own), both standards and decisions
live in the Standards directory. The reason is that this is the
canonical place to look for standardization relevant things and
spreading it to two directories makes both searching and numbering
more difficult.

Add a clarifying README.md and fix top level README.md

Signed-off-by: Kurt Garloff <kurt@garloff.de>