Renumber metering 0110/0111 to 0410/0411.
This put it into the Ops track -- the metering approach is not
restricted to the IaaS layer, even if we implement it there first.

This also avoids the conflict with the previously existent
scs-0110-v1-ssd-flavors.
(Note that we don't want to have double numbers, not even between
decisions and standards.)

Signed-off-by: Kurt Garloff <kurt@garloff.de>