Adds a new decision record about the implementation of sonobuoy as KaaS test suite (#392)
* Inital sonobuoy test suite

* provide python example
+ provide go example

Signed-off-by: Toni Finger <toni.finger@cloudandheat.com>

* Create Decision Record on the ussage of Sonobuoy

Add an additional 'kaas-sonobuoy-go-example-e2e-framework'
as an example of using the e2e framework provided by kubernetes-sigs/

Signed-off-by: Toni Finger <toni.finger@cloudandheat.com>

* Add proof of work for sonobuoy python plugin

With the merge of #360 a proof of work for the implementation of a
sonobuoy python plugin is provided.
Therefor it is now linked to the decision record.

Signed-off-by: Toni Finger <toni.finger@cloudandheat.com>

* Implement review changes

* fix sonobuoy e2e framework example
* update Standards/scs-0200-v1-sonobuoy-kaas-conformance-tests.md
* apply code review suggestions
* fix markdown lint errors

Co-authored-by: Hannes Baum <hannes.baum@cloudandheat.com>
Signed-off-by: Toni Finger <toni.finger@cloudandheat.com>

* Remove second e2e-framework example

Signed-off-by: Toni Finger <toni.finger@cloudandheat.com>

* Remove realtive link paths

Relative link paths are not allowed in Markdown documents.  Links got moved to backticks

Signed-off-by: Toni Finger <toni.finger@cloudandheat.com>

* Restructure Markdown

Restructure raw content of the decision record to be better readable
in raw format.

Signed-off-by: Toni Finger <toni.finger@cloudandheat.com>

* Add first draft decision

As a conclusion of the investigation and an aid for the decision-making process
this commit adds a first decsion on how to proceed with scs-0200-v1-sonobuoy-kaas-conformance-tests

Signed-off-by: Toni Finger <toni.finger@cloudandheat.com>

* Update scs-0200-v1-sonobuoy-kaas-conformance-tests.md

just fixed a few typos here and there. And i changed PoW (https://en.wikipedia.org/wiki/Proof_of_work) to PoC https://en.wikipedia.org/wiki/Proof_of_concept :)

Signed-off-by: Malte Münch <muench@b1-systems.de>

* Restructure e2e-framework

To give a suggestion on how to structure/name tests using the e2e-framework
I restructured the code example.

Signed-off-by: Toni Finger <toni.finger@cloudandheat.com>

* Update structure, grammar and tabs (#345)

Most of the documents needed small updates either for text flow, grammar, tabs or structure.

Signed-off-by: Hannes Baum <hannes.baum@cloudandheat.com>

---------

Signed-off-by: Toni Finger <toni.finger@cloudandheat.com>
Signed-off-by: Malte Münch <muench@b1-systems.de>
Signed-off-by: Hannes Baum <hannes.baum@cloudandheat.com>
Co-authored-by: Hannes Baum <hannes.baum@cloudandheat.com>
Co-authored-by: Malte Münch <muench@b1-systems.de>