More verbose output: Report passed tests, tool invocations.
When running the test suite interactively, I want to see know what it's
currently doing and what the current state of success is.
This is actually already logged, but with log level debug and we have
no way to make this visible by cmd line options. Rather than introducing
an option '-d|--debug', let's just attach it to '-v|--verbose', as it
does not add all that much output.

Also, when used with '-v', let's also report all passed checks in the
summary.

Signed-off-by: Kurt Garloff <kurt@garloff.de>