More specific meaning of GPU h modifiers.
h on the SMs/CUs/EUs: High frequency
h on the VRAM: High bandwidth

Again, this is really only to differentiate if a vendor has several
otherwise similar models that have a material difference in frequency
or bandwidth, such as e.g. a GDDR6 vs an HBM2e veriant ...
or a low-power, low-frequency variant.

Signed-off-by: Kurt Garloff <kurt@garloff.de>