Increase timeout to 80s. Backoff poll frequency.
Timeout was 60s, increase to 80s. This may help on heavily loaded
storage environments.

We see ~60 debug messages for each timed out wait. This is more than
needed and puts a bit more load on the control plane than needed.
So change the approach: Start with 0.5s polling freq but increase
wait time by 0.1s each time, slowly backing off. So we produce only
35 debug lines and API calls before running into the timeout.

Sidenote: Exponential backoff is a well-known approach to deal
with congestion. We kept it simple (linear back-off), as our timeout
is not huge.

Signed-off-by: Kurt Garloff <kurt@garloff.de>