Add os_purpose field.
Also add a link to OpenStack specs for os_distro and explain why we
need it.

This is patch 1 from several. It adds this os_purpose field as
*recommended*. As it does not break certification, this can be done in
the existing v1 of the scs-0102 standard. I suggest we discuss this one
first until we are aligned. In parallel, I'd like to prepare a blog
article that serves as user guide how to select images in a portable way
without relying on the name. We can reference it in tne implementation
notes. We should also enhance the image-manager input to carry this
field soon.

In a future patch, we create a v2 of scs-0102 which *requires* the
os_purpose field. The naming in scs-0102 is already a recommendation
only, so that can remain. scs-0104 will need a new version then,
as it does use the recommended names as mandatory for the mandatory
images.

Signed-off-by: Kurt Garloff <kurt@garloff.de>