Remove Status field from Supplement docs
This field causes some confusion for Supplement docs, for the
standard scs-0001-v1 merely states that these docs may be kept
in Draft state, which provides no certainty as to what should
be done. Here, I propose to go for maximum certainty by simply
stating they are kept in Draft state in perpetuity, and
therefore, the field is not even present in them.

Signed-off-by: Matthias Büchse <matthias.buechse@alasca.cloud>